qa/requirements.txt
behave==1.2.5
locustio==0.7.5
PyHamcrest==1.9.0
python-owasp-zap-v2.4==0.0.11
selenium==3.4.3
